blur_app/requirements.txt
numpy
numba
matplotlib
pillow
requests